cgroups: Allow memory cgroup support to be included but disabled
Memory cgroup support has some run-time overhead, so it's useful to
include it in a distribution kernel without enabling it by default.
Add a kernel config option to disable it by default and a kernel
parameter 'cgroup_enable' as the opposite to 'cgroup_disable'.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>

Gbp-Pq: Topic features/all
Gbp-Pq: Name cgroups-Allow-memory-cgroup-support-to-be-included-b.patch