line6: fix memory leaks in line6_init_midi()
commit 982d6ab50115025c63f4b9f7175614008c41d449 upstream.

If the first call to line6_midibuf_init() fails we'll leak a little
bit of memory. If the second call fails we'll leak a bit more. This
happens when we return from the function and the local variable
'line6midi' goes out of scope.

Signed-off-by: Jesper Juhl <jj@chaosbits.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>

Gbp-Pq: Topic features/all/line6
Gbp-Pq: Name 0001-line6-fix-memory-leaks-in-line6_init_midi.patch