Input: wacom - cleanup feature report for bamboos
commit 6e8ec5379c7ab9a57190e23af173aee74f88e54a upstream.

Only the stylus interface on Bamboo's has a feature report that can
be used to set to wacom mode.  The touch interface only has 1 report mode
and will return errors for those get/sets requests.

The get request was always erroring out because it was not marked as
an input request. Only down side of error was needlessly resending the
set request 5 times.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Acked-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Dmitry Torokhov <dtor@mail.ru>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Gbp-Pq: Topic features/all/wacom
Gbp-Pq: Name 0001-Input-wacom-cleanup-feature-report-for-bamboos.patch