alpha: add io{read,write}{16,32}be functions
commit 25534eb7707821b796fd84f7115367e02f36aa60 upstream.

These functions are used in some PCI drivers with big-endian
MMIO space.

Admittedly it is almost certain that no one this side of the
Moon would use such a card in an Alpha but it does get us
closer to being able to build allyesconfig or allmodconfig,
and it enables the Debian default generic config to build.

Signed-off-by: Michael Cree <mcree@orcon.net.nz>

Gbp-Pq: Topic bugfix/alpha
Gbp-Pq: Name alpha-add-io-read-write-16-32-be-functions.patch