ARM: Kirkwood: Add configuration for MPP12 as GPIO
commit 527ef0550d79e3b3a0ef8f5061072075afef6aaf upstream.

The MPP12 is listed in the 6281 HW manual as output only but the iconnect
board from iomega is using it as GPIO (there's a button connected on it). So,
I'm adding a definition for the MPP12 as GPIO. As I've no informations about
this and which kirkwood are "affected", I'm adding a new #define instead of
modifying the current one for MPP12.

Signed-off-by: Arnaud Patard <arnaud.patard@rtp-net.org>
Signed-off-by: Nicolas Pitre <nico@fluxnic.net>


Gbp-Pq: Topic features/arm
Gbp-Pq: Name kirkwood-add-configuration-for-mpp12-as-gpio.patch