Staging: hv: storvsc: Disable clustering
commit 039db52de9c5682ee10a3cd69f2d76db40b1dee0 upstream.

Disable clustering, since the host side on Hyper-V requires that
each I/O element not exceed the page size. As part of this
cleanup, get rid of the function to merge bvecs, as the primary
reason for this function was to avoid having an element exceed
the page size.

Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
Signed-off-by: Haiyang Zhang <haiyangz@microsoft.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>

Gbp-Pq: Topic features/x86/hyperv
Gbp-Pq: Name 0021-Staging-hv-storvsc-Disable-clustering.patch