driver core: remove __must_check from device_create_file
commit b9d4e714a86a4e88c2f530c76597f7025e5851d6 upstream.

With the conversion of the sysdev to a real struct device, more drivers
are calling device_create_file, and some of them don't check the return
value, which isn't wise.

But as they happen to be in parts of the kernel where a warning is
considered an error (i.e. powerpc), this breaks the build.  So for now,
remove the marking on the function, which fixes the build problems.

Reported-by: Stephen Rothwell <sfr@canb.auug.org.au>
Cc: Kay Sievers <kay.sievers@vrfy.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>

Gbp-Pq: Topic features/all/cpu-devices
Gbp-Pq: Name driver-core-remove-__must_check-from-device_create_f.patch