be2net: add descriptions for stat counters reported via ethtool
commit d45b9d39a1aed7851948460d29b843ce70eb0a68 upstream.

Also rename a few counters appropritely and delete 2 counters that are not
implemented in HW.

vlan_mismatch_drops does not exist in BE3 and is accounted for in
address_mismatch_drops. Do the same thing for BE2 and Lancer.

Signed-off-by: Sathya Perla <sathya.perla@emulex.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Gbp-Pq: Topic features/all/be2net
Gbp-Pq: Name 0029-be2net-add-descriptions-for-stat-counters-reported-v.patch