driver core: cpu: remove kernel warning when removing a cpu
commit 2885e25c422fb68208f677f944a45fce8eda2a3c upstream.

With the movement of the cpu sysdev code to be real stuct devices, now
when we remove a cpu from the system, the driver core rightfully
complains that there is not a release method for this device.

For now, paper over this issue by quieting the driver core, but comment
this in detail.  This will be resolved in future kernels to be solved
properly.

Reported-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Tested-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
[bwh: Adjust context for textual conflict with commit
 fad12ac8c8c2591c7f4e61d19b6a9d76cd49fafa]

Gbp-Pq: Topic features/all/cpu-devices
Gbp-Pq: Name driver-core-cpu-remove-kernel-warning-when-removing-.patch