virtio-scsi: SCSI driver for QEMU based virtual machines
commit 4fe74b1cb051dc9d47a80e263c388cf1651783d4 upstream.

The virtio-scsi HBA is the basis of an alternative storage stack
for QEMU-based virtual machines (including KVM).  Compared to
virtio-blk it is more scalable, because it supports many LUNs
on a single PCI slot), more powerful (it more easily supports
passthrough of host devices to the guest) and more easily
extensible (new SCSI features implemented by QEMU should not
require updating the driver in the guest).

Acked-by: Rusty Russell <rusty@rustcorp.com.au>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
[bwh: Backported to 3.2:
 - s/virtqueue_add_buf/&_gfp/
 - Remove virtio driver PM operations]

Gbp-Pq: Topic features/all/virtio_scsi
Gbp-Pq: Name 0001-SCSI-virtio-scsi-SCSI-driver-for-QEMU-based-virtual-.patch