driver-core: cpu: fix kobject warning when hotplugging a cpu
commit 29bb5d4fd3140a7d5d02d858118c74a45f15c296 upstream.

Due to the sysdev conversion to struct device, the cpu objects get
reused when adding a cpu after offlining it, which causes a big warning
that the kobject portion is not properly initialized.

So clear out the object before we register it again, so all is quiet.

Reported-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Tested-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Gbp-Pq: Topic features/all/cpu-devices
Gbp-Pq: Name driver-core-cpu-fix-kobject-warning-when-hotplugging.patch