Staging: hv: storvsc: Move the storage driver out of the staging area
commit 89ae7d709357df3a60e81c57613859d91d093ace upstream.

The storage driver (storvsc_drv.c) handles all block storage devices
assigned to Linux guests hosted on Hyper-V. This driver has been in the
staging tree for a while and this patch moves it out of the staging area.

James was willing to apply this patch during the 3.3-rc phase and a decision
was taken to defer this to 3.4 since Greg had queued up a bunch of storvsc
patches for 3.4. Now that Greg has applied all of the pending storvsc patches,
I am sending this patch to move this driver out of staging. Based on James'
recommendation, this patch gets rid of the unneeded files in the staging/hv
directory.

Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
Acked-by: James Bottomley <JBottomley@Parallels.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
[bwh: Adjusted to apply after backported commits 5c1b10ab7f93
 '[SCSI] storvsc: Account for in-transit packets in the RESET path' and
 9d2696e658ef '[SCSI] storvsc: Initialize the sglist']

Gbp-Pq: Topic features/x86/hyperv
Gbp-Pq: Name 0067-Staging-hv-storvsc-Move-the-storage-driver-out-of-th.patch