ARM: kirkwood: fdt: define uart[01] as disabled, enable uart0
commit 163f2cea673a4ae831ad2cd26d8f01977c3add93 upstream.

Define both uarts in kirkwood.dtsi as they are common to all kirkwood
SoCs.  Each board may enable all or none of them, so they are disabled
by default.  uart0 is enabled for the dreamplug.

tclk can vary for each board, so we leave it undefined in the kirkwood
dtsi.  Each board can then set it as appropriate when enabling the uart.

Signed-off-by: Jason Cooper <jason@lakedaemon.net>


Gbp-Pq: Topic features/arm
Gbp-Pq: Name kirkwood-fdt-define-uart01-as-disabled.patch