Tools: hv: Fully support the new KVP verbs in the user level daemon
commit db425334e5bb7fa65bbbd7bea9d79842f65bcf45 upstream.

Now fully support the new KVP messages in the user level daemon. Hyper-V defines
multiple persistent pools to which the host can write/read/modify KVP tuples.
In this patch we implement a file for each specified pool, where the KVP tuples
will be stored in the guest.

Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
Reviewed-by: Haiyang Zhang <haiyangz@microsoft.com>
Reviewed-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Gbp-Pq: Topic features/x86/hyperv
Gbp-Pq: Name 0074-Tools-hv-Fully-support-the-new-KVP-verbs-in-the-user.patch